added yes or no question when deleting a drug or user

+++ Med Match/Forms/ShowDrugs.cs
@@ -146,16 +146,29 @@ namespace Med_Match.Forms
 
         private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
-            if (e.ColumnIndex == dataGridView1.Columns["Delete"].Index)
+            try
             {
-                var documentId = dataGridView1.Rows[e.RowIndex].Cells["_Id"].Value.ToString();
-                var drugName = dataGridView1.Rows[e.RowIndex].Cells["name"].Value.ToString();
-                var filter = Builders<BsonDocument>.Filter.Eq("_id", ObjectId.Parse(documentId));
-                collection.DeleteOne(filter);
-                MessageBox.Show(drugName +  " Deleted Successfully");
-                search_btn.PerformClick();
+                if (e.ColumnIndex == dataGridView1.Columns["Delete"].Index)
+                {
+                    var documentId = dataGridView1.Rows[e.RowIndex].Cells["_Id"].Value.ToString();
+                    var drugName = dataGridView1.Rows[e.RowIndex].Cells["name"].Value.ToString();
+                    var filter = Builders<BsonDocument>.Filter.Eq("_id", ObjectId.Parse(documentId));
+                    DialogResult dr = MessageBox.Show("Do You Want To Delete This Drug?", "Delete", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                    if (dr == DialogResult.Yes)
+                    {
+                        collection.DeleteOne(filter);
+                        MessageBox.Show(drugName + " Deleted Successfully");
+                        search_btn.PerformClick();
+                    }
 
+
+                }
             }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+            
         }
     }
 }
